Move public APIs to single index.ts files
