+++ packages/admin-react-select/src/index.ts
@@ -1,4 +1,9 @@
-export * from "./ReactSelect";
-export * from "./FinalFormReactSelect";
-export * from "./FinalFormReactSelectStaticOptions";
+export { CometAdminSelectProps, ControlInput, ReactSelect, ReactSelectAsync, ReactSelectAsyncCreatable, ReactSelectCreatable } from "./ReactSelect";
+export {
+    FinalFormReactSelect,
+    FinalFormReactSelectAsync,
+    FinalFormReactSelectAsyncCreatable,
+    FinalFormReactSelectCreatable,
+} from "./FinalFormReactSelect";
+export { FinalFormReactSelectStaticOptions } from "./FinalFormReactSelectStaticOptions";
 export { CometAdminSelectClassKeys } from "./ReactSelect.styles";
