--- packages/admin-rte/src/core/Controls/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Controls";
