--- packages/admin-rte/src/core/extension/Link/index.ts
@@ -1,2 +1,0 @@
-export { default as ToolbarButton } from "./ToolbarButton";
-export { default as Decorator } from "./Decorator";
