--- packages/admin-rte/src/core/extension/LinksRemove/index.ts
@@ -1,1 +1,0 @@
-export { default as ToolbarButton } from "./ToolbarButton";
