+++ packages/admin-stories/src/vendor.d.ts
