--- packages/admin/src/common/index.ts
@@ -1,2 +1,0 @@
-export * from "./ClearInputButton";
-export * from "./toolbar";
