+++ packages/admin/src/common/toolbar/automatictitleitem/ToolbarAutomaticTitleItem.tsx
@@ -1,7 +1,7 @@
 import { Typography } from "@material-ui/core";
 import * as React from "react";
 
-import { useStackApi } from "../../../stack";
+import { useStackApi } from "../../../stack/Api";
 import { ToolbarItem } from "../item/ToolbarItem";
 import { useThemeProps } from "./ToolbarAutomaticTitleItem.styles";
