--- packages/admin/src/common/toolbar/index.ts
@@ -1,16 +1,0 @@
-export * from "./Toolbar";
-export { CometAdminToolbarClassKeys, ToolbarThemeProps } from "./Toolbar.styles";
-export * from "./item/ToolbarItem";
-export { CometAdminToolbarItemClassKeys } from "./item/ToolbarItem.styles";
-export * from "./automatictitleitem/ToolbarAutomaticTitleItem";
-export { ToolbarAutomaticTitleItemThemeProps } from "./automatictitleitem/ToolbarAutomaticTitleItem.styles";
-export * from "./titleitem/ToolbarTitleItem";
-export { ToolbarTitleItemThemeProps } from "./titleitem/ToolbarTitleItem.styles";
-export * from "./breadcrumb/ToolbarBreadcrumbs";
-export { ToolbarBreadcrumbsThemeProps, CometAdminToolbarBreadcrumbsClassKeys } from "./breadcrumb/ToolbarBreadcrumbs.styles";
-export * from "./backbutton/ToolbarBackButton";
-export * from "./actions/ToolbarActions";
-export { CometAdminToolbarActionsClassKeys } from "./actions/ToolbarActions.styles";
-export * from "./fillspace/ToolbarFillSpace";
-export { CometAdminToolbarFillSpaceClassKeys } from "./fillspace/ToolbarFillSpace.styles";
-export { CometAdminToolbarBackButtonClassKeys, ToolbarBackButtonThemeProps } from "./backbutton/ToolbarBackButton.styles";
