--- packages/admin/src/error/errorboundary/index.ts
@@ -1,2 +1,0 @@
-export * from "./ErrorBoundary";
-export * from "./RouteWithErrorBoundary";
