--- packages/admin/src/error/errordialog/index.ts
@@ -1,6 +1,0 @@
-export * from "./ErrorDialogContext";
-export * from "./ErrorDialog";
-export * from "./ErrorDialogProvider";
-export * from "./useErrorDialog";
-export * from "./ErrorScope";
-export * from "./createErrorDialogApolloLink";
