--- packages/admin/src/error/index.ts
@@ -1,2 +1,0 @@
-export * from "./errorboundary";
-export * from "./errordialog";
