--- packages/admin/src/fetchProvider/index.ts
@@ -1,1 +1,0 @@
-export * from "./fetch";
