--- packages/admin/src/fileIcons/index.ts
@@ -1,1 +1,0 @@
-export * from "./FileIcon";
