--- packages/admin/src/form/index.ts
@@ -1,16 +1,0 @@
-export * from "./Field";
-export * from "./FieldContainer";
-export * from "./FormPaper";
-export * from "./FormSection";
-export * from "./Input";
-export * from "./InputBase";
-export * from "./FinalFormInput";
-export * from "./Checkbox";
-export * from "./Radio";
-export * from "./Switch";
-export * from "./TextField";
-export * from "./Select";
-export * from "./FinalFormSelect";
-export * from "./FinalFormRangeInput";
-export * from "./FinalFormSearchTextField";
-export * from "./FinalFormSearchTextField.styles";
