--- packages/admin/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./useWindowSize";
