--- packages/admin/src/mui/index.ts
@@ -1,4 +1,0 @@
-export * from "./menu";
-export * from "./ThemeProvider";
-export * from "./MasterLayout";
-export * from "./MainContent";
