--- packages/admin/src/mui/menu/index.ts
@@ -1,7 +1,0 @@
-export { MenuItemRouterLink } from "./ItemRouterLink";
-export * from "./CollapsibleItem";
-export * from "./Item";
-export * from "./Menu";
-export * from "./ItemAnchorLink";
-export { CometAdminMenuClassKeys } from "./Menu.styles";
-export * from "./Context";
