--- packages/admin/src/router/index.ts
@@ -1,5 +1,0 @@
-export * from "./BrowserRouter";
-export * from "./ConfirmationDialog";
-export * from "./Context";
-export * from "./Prompt";
-export * from "./PromptHandler";
