--- packages/admin/src/stack/backbutton/index.ts
@@ -1,2 +1,0 @@
-export * from "./StackBackButton";
-export { CometAdminStackBackButtonThemeProps } from "./StackBackButton.styles";
