--- packages/admin/src/stack/breadcrumbs/index.ts
@@ -1,2 +1,0 @@
-export * from "./StackBreadcrumbs";
-export { CometAdminStackBreadcrumbsClassKeys, CometAdminStackBreadcrumbsThemeProps } from "./StackBreadcrumbs.styles";
