--- packages/admin/src/stack/index.ts
@@ -1,9 +1,0 @@
-export * from "./Api";
-export * from "./Breadcrumb";
-export * from "./Page";
-export { Stack } from "./Stack";
-export * from "./StackPageTitle";
-export * from "./Switch";
-export * from "./SwitchMeta";
-export * from "./breadcrumbs";
-export * from "./backbutton";
