--- packages/admin/src/table/index.ts
@@ -1,25 +1,0 @@
-// Export folders
-export * from "./paging";
-export * from "./excelexport";
-
-// Export files
-export * from "./AddButton";
-export * from "./DeleteButton";
-export * from "./ExcelExportButton";
-export * from "./LocalChangesToolbar";
-export * from "./Pagination";
-export * from "./Table";
-export * from "./TableDndOrder";
-export * from "./TableFilterFinalForm";
-export * from "./TableLocalChanges";
-export * from "./TableQuery";
-export * from "./TableQueryContext";
-
-export * from "./useTableQuery";
-export * from "./useTableQuerySort";
-export * from "./useTableQueryPaging";
-export * from "./useTableQueryFilter";
-export * from "./usePersistedStateId";
-export * from "./usePersistedState";
-export * from "./withDirtyHandlerApi";
-export * from "./withTableQueryContext";
