--- packages/admin/src/table/paging/index.ts
@@ -1,6 +1,0 @@
-export * from "./createRelayPagingActions";
-export * from "./createRestPagingActions";
-export * from "./createPagePagingActions";
-export * from "./createRestStartLimitPagingActions";
-
-export * from "./IPagingInfo";
