+++ packages/admin/src/themeAugmentation.ts
@@ -1,22 +1,15 @@
-import { ClearInputButtonThemeProps, CometAdminClearInputButtonClassKeys } from "./common";
-import { CometAdminErrorBoundaryClassKeys, ErrorBoundaryThemeProps } from "./error";
-import {
-    CometAdminFinalFormRangeInputClassKeys,
-    CometAdminFormFieldContainerClassKeys,
-    CometAdminFormPaperKeys,
-    CometAdminFormSectionKeys,
-    CometAdminInputBaseClassKeys,
-    FieldContainerThemeProps,
-} from "./form";
-import {
-    CometAdminMasterLayoutClassKeys,
-    CometAdminMenuClassKeys,
-    CometAdminMenuCollapsibleItemClassKeys,
-    CometAdminMenuItemClassKeys,
-    MasterLayoutThemeProps,
-    MenuCollapsibleItemThemeProps,
-    MenuThemeProps,
-} from "./mui";
+import { ClearInputButtonThemeProps, CometAdminClearInputButtonClassKeys } from "./common/ClearInputButton";
+import { CometAdminErrorBoundaryClassKeys, ErrorBoundaryThemeProps } from "./error/errorboundary/ErrorBoundary";
+import { CometAdminFormFieldContainerClassKeys, FieldContainerThemeProps } from "./form/FieldContainer";
+import { CometAdminFinalFormRangeInputClassKeys } from "./form/FinalFormRangeInput";
+import { CometAdminFormPaperKeys } from "./form/FormPaper";
+import { CometAdminFormSectionKeys } from "./form/FormSection";
+import { CometAdminInputBaseClassKeys } from "./form/InputBase";
+import { CometAdminMasterLayoutClassKeys, MasterLayoutThemeProps } from "./mui/MasterLayout";
+import { CometAdminMenuCollapsibleItemClassKeys, MenuCollapsibleItemThemeProps } from "./mui/menu/CollapsibleItem";
+import { CometAdminMenuItemClassKeys } from "./mui/menu/Item";
+import { MenuThemeProps } from "./mui/menu/Menu";
+import { CometAdminMenuClassKeys } from "./mui/menu/Menu.styles";
 
 declare module "@material-ui/core/styles/overrides" {
     interface ComponentNameToClassKey {
